# api/openapi.yaml
openapi: 3.1.0
info:
  title: vks4 Control Plane API
  version: 0.1.0
  description: |
    REST API сервиса control-plane: аутентификация, RBAC, комнаты,
    участники, layout, записи, пользователи, аудит, метрики.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /rooms:
    get:
      tags: [rooms]
      parameters:
        - { in: query, name: q,      schema: { type: string } }
        - { in: query, name: limit,  schema: { type: integer, default: 50, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Room" } }
                  total: { type: integer }
    post:
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomCreate" }
      responses:
        "201":
          description: created
          content: { application/json: { schema: { $ref: "#/components/schemas/Room" } } }
  /rooms/{roomId}:
    parameters:
      - { in: path, name: roomId, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [rooms]
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Room" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomUpdate" }
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Room" } } }
    delete:
      tags: [rooms]
      responses:
        "204": { description: deleted }
...
  /rooms/{roomId}/layout:
    parameters: [ { in: path, name: roomId, required: true, schema: { type: string, format: uuid } } ]
    put:
      tags: [layouts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Layout" }
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Layout" } } }
...
  /rooms/{roomId}/recording:
    parameters: [ { in: path, name: roomId, required: true, schema: { type: string, format: uuid } } ]
    post:
      tags: [recordings]
      summary: start
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Recording" } } }
    delete:
      tags: [recordings]
      summary: stop
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Recording" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    RoomMode:
      type: string
      enum: [meeting, lecture, webinar]
    Room:
      type: object
      required: [id, name, mode, createdAt, locked, recording, defaultLayout]
      properties:
        id:             { type: string, format: uuid }
        name:           { type: string }
        description:    { type: string }
        mode:           { $ref: "#/components/schemas/RoomMode" }
        maxParticipants: { type: integer, default: 50 }
        waitingRoom:    { type: boolean, default: false }
        locked:         { type: boolean }
        recording:      { type: boolean }
        defaultLayout:  { $ref: "#/components/schemas/Layout" }
        joinUrl:        { type: string }
        createdAt:      { type: string, format: date-time }
        createdBy:      { type: string, format: uuid }
    RoomCreate:
      type: object
      required: [name]
      properties:
        name:            { type: string }
        description:     { type: string }
        mode:            { $ref: "#/components/schemas/RoomMode", default: meeting }
        maxParticipants: { type: integer, default: 50 }
        waitingRoom:     { type: boolean, default: false }
        defaultLayout:   { $ref: "#/components/schemas/Layout" }
    RoomUpdate:
      type: object
      properties:
        name:            { type: string }
        description:     { type: string }
        mode:            { $ref: "#/components/schemas/RoomMode" }
        maxParticipants: { type: integer }
        waitingRoom:     { type: boolean }
        defaultLayout:   { $ref: "#/components/schemas/Layout" }
    LayoutMode:
      type: string
      enum: [grid, active_speaker, lecture, pip, custom]
    LayoutCell:
      type: object
      required: [id, x, y, w, h]
      properties:
        id:     { type: string, description: "peerId или 'speaker'/'presenter'" }
        x:      { type: number }
        y:      { type: number }
        w:      { type: number }
        h:      { type: number }
        zIndex: { type: integer, default: 0 }
    Layout:
      type: object
      required: [mode]
      properties:
        mode:    { $ref: "#/components/schemas/LayoutMode" }
        width:   { type: integer, default: 1280 }
        height:  { type: integer, default: 720 }
        background:
          type: object
          properties:
            color: { type: string, default: "#000000" }
            url:   { type: string }
        logo:
          type: object
          properties:
            url: { type: string }
            x:   { type: number }
            y:   { type: number }
            w:   { type: number }
            h:   { type: number }
        cells:
          type: array
          items: { $ref: "#/components/schemas/LayoutCell" }
...
    Recording:
      type: object
      required: [id, roomId, status, startedAt]
      properties:
        id:        { type: string, format: uuid }
        roomId:    { type: string, format: uuid }
        status:    { type: string, enum: [running, finished, failed] }
        startedAt: { type: string, format: date-time }
        endedAt:   { type: string, format: date-time, nullable: true }
        sizeBytes: { type: integer, nullable: true }
        url:       { type: string, nullable: true }
...